**refactor(extractor): remove `BaseExtractor` utility and related modules**
### Details:
- Removed the entire `BaseExtractor` implementation (`common.rs`) from the `rdlp-extractor` crate.
  - Included security constants, static selectors, regex patterns, and shared utility functionality.
  - Extraction logic such as `fetch_webpage`, `extract_id_from_url`, and format-building utilities were retired.
- Cleaned up outdated, unused shared utilities:
  - URL validation (`validate_url_security`, `is_private_host`).
  - Metadata, title, and description multi-strategy extractors.
  - Filesize detection logic using HEAD requests and range fallbacks.
  - String truncation, log utilities, regex-based quality parsing, and dynamic FFmpeg format generation.

### Impact:
Simplifies the codebase by removing redundant or legacy infrastructure that no longer aligns with the new extraction framework. Ensures maintainability by deprecating unused logic and reducing potential attack surface area.

### Testing:
Verified successful builds and functionality after removal. Confirmed that no existing extractors relied on the removed utilities.

### Notes:
This is a breaking change. Existing implementations depending on `BaseExtractor` must be refactored to use the updated modular helpers or implement site-specific solutions.

**BREAKING CHANGE:** `BaseExtractor` and all related utilities are no longer available. Developers should migrate to newer extraction modules or implement custom logic as needed.
